MLflow/mlflow-docker/requirements.txt
boto3==1.26.38
numpy==1.23.5
mlflow==2.2.1
pandas==1.5.2
scikit-learn==1.2.1
scikit-optimize==0.9.0
pymysql==1.0.2
python-dotenv==0.20.0